requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.5.0
twilio==9.3.7
Pillow==10.4.0
python-decouple==3.8
django-crispy-forms==2.3
crispy-bootstrap4==2024.1
django-bootstrap4==24.4
reportlab==4.2.5
openpyxl==3.1.5
celery==5.4.0
redis==5.2.0
django-celery-beat==2.7.0
python-dateutil==2.9.0